Refactor to remove the separate profile editing hook
